Final Version

+++ src/com/dao/Dao.java
@@ -3,6 +3,7 @@ package com.dao;
 import com.bean.Faculty;
 import com.dbUtil.DBUtil;
 
+import javax.swing.table.TableRowSorter;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -13,79 +14,91 @@ public class Dao {
     public boolean addFaculty(Faculty f){
         Connection con= DBUtil.getConnection();
         PreparedStatement pst=null;
-        String query=null;
+        String query;
+        boolean flag=true;
         try{
             query="Insert into Faculty values(Faculty_Sequence.nextval,?,?,?,?,?,?,?)";
             pst=con.prepareStatement(query);
-            pst.setInt(1,f.getId());
-            pst.setString(2,f.getName());
-            pst.setString(3,f.getAddress());
-            pst.setString(4,f.getStream());
-            pst.setDouble(5,f.getSalary());
-            pst.setInt(6,f.getYears_of_experience());
-            pst.setLong(7,f.getPhoneno());
+            pst.setString(1,f.getName());
+            pst.setString(2,f.getAddress());
+            pst.setString(3,f.getStream());
+            pst.setDouble(4,f.getSalary());
+            pst.setInt(5,f.getYears_of_experience());
+            pst.setLong(6,f.getPhoneno());
             pst.setString(7,f.getGrade());
             pst.executeUpdate();
         }catch(SQLException e){
+            flag=false;
             e.printStackTrace();
         }
         finally {
             try {
                 pst.close();
-            } catch (SQLException e) {
+            } catch (SQLException|NullPointerException e) {
                 e.printStackTrace();
             }
             DBUtil.closeConnection(con);
-            return true;
         }
+        return flag;
     }
 
 
 
-    public boolean removeFaculty(int id){
+    public boolean removeFaculty(Integer id){
+        boolean isValid=isFacultyValid(id);
+        if(!isValid)
+            return false;
+
         Connection con=DBUtil.getConnection();
         PreparedStatement pst = null;
-        String query = null;
+        String query;
+        boolean flag=true;
         try{
             query="Delete from Faculty where id="+id;
             pst=con.prepareStatement(query);
             pst.executeUpdate();
         }catch(SQLException e){
+            flag=false;
             e.printStackTrace();
-        }
-        finally {
+        }finally {
             try {
                 pst.close();
-            } catch (SQLException e) {
+            } catch (SQLException|NullPointerException e) {
                 e.printStackTrace();
             }
             DBUtil.closeConnection(con);
-            return true;
         }
+        return flag;
     }
 
 
 
-    public boolean updateFaculty(int id, String address){
+    public boolean updateFaculty(Integer id, String address){
+        boolean isValid=isFacultyValid(id);
+        if(!isValid)
+            return false;
+
         Connection con=DBUtil.getConnection();
         PreparedStatement pst=null;
         String query=null;
+        boolean flag= true;
         try{
-            query="Update Faculty set address="+address+" where id="+id;
+            query="Update Faculty set address='"+address+"' where id="+id;
             pst=con.prepareStatement(query);
             pst.executeUpdate();
         }catch(SQLException e){
+            flag=false;
             e.printStackTrace();
         }
         finally {
             try {
                 pst.close();
-            } catch (SQLException e) {
+            } catch (SQLException|NullPointerException e) {
                 e.printStackTrace();
             }
             DBUtil.closeConnection(con);
-            return true;
         }
+        return flag;
     }
 
 
@@ -119,11 +132,41 @@ public class Dao {
         finally {
             try{
                 pst.close();
-            }catch(SQLException e){
+            }catch(SQLException|NullPointerException e){
                 e.printStackTrace();
             }
             DBUtil.closeConnection(con);
-            return facs;
         }
+        return facs;
     }
+
+    public boolean isFacultyValid(Integer id){
+        Connection con=DBUtil.getConnection();
+        PreparedStatement pst=null;
+        ResultSet rs=null;
+        String query;
+        boolean flag=true;
+        try{
+            query="Select * from Faculty where id="+id;
+            pst=con.prepareStatement(query);
+            rs=pst.executeQuery();
+            int count=0;
+            while(rs.next()) count++;
+            if(count==0)
+                flag=false;
+        }catch(SQLException e){
+            flag=false;
+            e.printStackTrace();
+        }
+        finally {
+            try{
+                pst.close();
+            }catch(SQLException|NullPointerException e){
+                e.printStackTrace();
+            }
+            DBUtil.closeConnection(con);
+        }
+        return flag;
+    }
+
 }
